fix: image paths for deployed CM UI Docs website
There is no `/global/` segment in image paths on https://cm-ui-docs.saddleback.com
* Bad: https://cm-ui-docs.saddleback.com/global/images/marty-mcfly.jpg
* Good: https://cm-ui-docs.saddleback.com/images/marty-mcfly.jpg

Also fix ESLINT issues in `docs/src/dataDisplay/image.js` (including renaming it to have JSX extension and converting to a function component, because class component only had a `render()` method and nothing else.
