+++ docs/template.ejs
@@ -9,7 +9,7 @@
 
     <meta name="msapplication-TileColor" content="#084d7e">
     <meta name="msapplication-TileImage" content="/global/images/win8_svcc_logo_white.png">
-    <link rel="shortcut icon" href="/global/images/favicon.ico" type="image/x-icon">
+    <link rel="shortcut icon" href="images/favicon.ico" type="image/x-icon">
 </head>
 <body>
     <div id="coreApp"></div>
